requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
pydantic
requests
agno
memorisdk
openai
